Develop be

See merge request s09-final/S09P31B105!88

+++ server/user/src/main/java/com/wishme/user/domain/User.java
@@ -27,7 +27,7 @@ public class User {
     private String userNickname;
 
     @Column(name = "user_school_seq")
-    private Long userSchoolSeq;
+    private Integer userSchoolSeq;
 
     @Column(name = "uuid")
     private String uuid;
